requirements.txt
fastapi
uvicorn
sentence-transformers
chromadb
groq
python-dotenv
pypdf

#pip install -r requirements.txt